arm64: dts: beacon-imx8mn-kit: Fix PMIC node for 2019.04
The PMIC node is copied from the kernel, but it doesn't support
the proper 'compatible' flag, and therefore doesn't get properly
configured.

This patch sync's the device tree node for the PMIC with that of
the i.MX8M Mini kit since it uses the same PMIC and LPDDR like
the Beacon Nano kit.

Signed-off-by: Adam Ford <aford173@gmail.com>