beacon_imx8mm: Enable GPIO Expanders
There are a few GPIO expanders.  One of them controls the
reset line to the USB host hub on the baseboard.  Without
this, the USB hub stays in reset and it cannot communicate
with devices on it.

This patch enables the GPIO expander driver to allow the
the hub to come out of reset.

Signed-off-by: Adam Ford <aford173@gmail.com>