board: beacon: imx8mm/n: Fix video enable on poweron
When the PWM1 pin is not pin-muxed, it defaults high which drives
the backlight on the LVDS screens, leaving a bright, white screen.

This patch attempts to pin-mux the PWM as soon as possible to
tie the GPIO1_IO01 pin to the PWM1 which is low when disabled.
This will then stay blank until Linux takes over the backlight.

Signed-off-by: Adam Ford <aford173@gmail.com>