board: beacon imx8m/n: Force MMC Environment to eMMC always
Currently, the environmental variables are stored on uSD when
booting from uSD or eMMC when booting from eMMC.  To preserve
the environmental variables with the SOM, this makes the device
return eMMC when the system queries the mmcdev location. To
fix this issue, there needs to be a check to determine if
mmcdev is set already.  This will restore booting from uSD if
desired.

Fixes: PHOENIX8-183

Signed-off-by: Adam Ford <aford173@gmail.com>