arm64: imx8mp: Resync device trees and fix misc bugs
The Ethernet reset is being removed by the device tree in U-Boot
before the device tree is passed to the kernel making the Ethernet
unavailable to Linux.  This could potentially be un-done if/when
the Ethernet is initialized in U-Boot.

Resync device trees.

Working:
microSD Booting
microSD @ SRD104
eMMC @ HS400-ES
USB Type-A ports via Hub
Ethernet on SOM

To-do:
TCPC configuration
USB-C (Ports J22 and J21)
Ethernet on baseboard

Signed-off-by: Adam Ford <aford173@gmail.com>