arm: dts: imx8mp-beacon: Fix FEC Ethernet resets
The phy-reset-gpios property in Linux is deprecated, but
it's still necessary for U-Boot, so the property is placed into
the -u-boot.dtsi file.  In the case of the imx8mp-beacon, this
reset was incorrect, so the Ethernet (FEC) on the baseboard was
not coming out of reset.  Fix this by setting the proper reset
gpio.