board: imx8mp_beacon: Fixup U-Boot Environment for MMC device
The imx8mp_beacon_defconfig configures CONFIG_ENV_IS_NOWHERE,
because the default is expecting the board to select the boot
device based on extra code that is run from late_board_init.
Since the environment variables will always be in eMMC, clearing
the CONFIG_ENV_IS_NOWHERE lets the system load from eMMC which
can be configured to set the boot device whereever the user
wants.  This also lets allows the removal of late_board_init
since it's not necessary any longer.