board: imx8mp_beacon: Setup UUID searching for Linux rootfs
Instead of specifying /dev/mmcblkXp2 for the root device,
this can be generated with a UUID which can be fetched by
specifying the mmcdev number desired to use for booting.
The advantage is that the user can boot from eMMC or SD
regardless of the boot source (QSPI, eMMC, microSD, etC).