freescale: TCPC: moving logs to debug
For the purposes of the imx8mp-beacon board, the TCPC controller
only needs to configure the power, and if it cannot configure
power, the beacon code already is configured to display an error.

Move the verbose logs to debug level which can be enabled
by setting DEBUG to squelch the extra messages.