ARM: imx: imx8mp_beacon: Replace distro boot command
The distro_boot command uses a series of entries to determine what
devices to load and and what order, but it's slow.  The other
Beacon devices use a singular mmcboot command which is dependent on
mmcdev being either 1 or 2 depending on whether or not the user
wants to boot from the baseboard microSD or the SOM eMMC.
Replace the distro_boot command with the standard to harmonize the
user experience with other devices and speed up the boot process.