regulator: bd718x7: Only bind children if PMIC_CHILDREN is enabled
In SPL, it may be necessary to access the PMIC. Due to limited
size, SPL may not use the whole regulator system, so instead it
manually writes a few register settings.

In order to enable bd718x7, the PMIC_CHILDREN flag needs to be set
which may overflow the code space using certain compilers.

To remove some overhead, make pmic_bind_children depend on PMIC_CHILDREN
so PMIC_CHILDREN can be disabled.