imx: imx8mn_beacon: Enable SPL_DM_PMIC_BD71837
In preparation for using the PMIC in SPL, the memory map needs to
change to match the NXP reference, the I2C pinmux needs to be
enabled in SPL, and and SPL_DM_PMIC_BD71837 must be enabled.

Fixing the memory map frees up space which will overflow without
this.