board: imx8mp-beacon: Fix Ethernet Controllers
The PHY on both the FEC and EQOS controllers on the imx8mp-beacon
board are connected to a Micrel PHY at address 3.  Enable the driver
for these PHY devices.

Remove the hard-coded PHY addrss from the imx8mp_beacon.h file and
fetch it from the device tree.

Properly configure the reset gpio, assertion, and deassertion under
the ethphy0 node instead of under the equos node.