imx: imx8mp_beacon: Disable mmcautodetect
For the imx8mp-beacon board, the eMMC always stores the environment
to the eMMC.  However, if mmcautodetect is enabled, the system
assumes that the 'mmcdev' variable should point to the same
device as the device holding the environment, but that isn't true.

Removing this lets the user select mmcdev based on whether they
want to load the Kernel and device tree from whatever device regardless
of what boot source or environment location is chosen.