arm: dts: imx8mp-beacon: Resync device tree from Kernel
The kernel has undergone some changes. Because the device tree sync is a
pull from Linux, there are some items missing which are not present
in the U-Boot imx8mp.dtsi file so they're removed.  Since the updated
name of the PMIC changed, the power_init_board() in spl.c needs
to be updated to search for the correct node in the device tree

Fixes: GRIF-83