imx: Introduce IMX_ENV_BOOT_DEFAULT config option
The current assumption is that the environment will be stored
on the device assoicated with 'mmcdev' if the environment is
stored in sd/mmc but the boot device isn't either of those.
This is hard-coded in arch/arm/mach-imx/mmc_env.c

However, for some boards, the desire may be to store the environment
based on the value of CONFIG_SYS_MMC_ENV_DEV regardless of the
value of mmcdev.  Add a config option, enabled by default which
can allow users to opt-out of this default feature.