arm: imx: imx8mp_beacon: Opt out of IMX_ENV_BOOT_DEFAULT
The environmental variables are store in eMMC, partition 2. When
booting from either microSD or eMMC, this isn't an issue.  However,
when booting from something other than sd/mmc, the default
behavior is to use the value of 'mmcdev' to determine where
the environment is stored.  This causes and issue when booting
from either USB or QSPI because the default mmcdev is 1.
In order to use eMMC (mmc dev 2) as the environment all the time,
disable IMX_ENV_BOOT_DEFAULT.

Fixes: GRIF-82