tcpc: Remove alert when no messages are found
The tcpc_pd_receive_message() returns a message and it's called
from a while statement which monitors all messages as long as there
are messages.  When no more messages are available, tcpc_pd_receive_message()
returns a value less than 1 indicating it's finshed and breaks out of the loop.
When there are no messages, an alert is shown on the terminal which looks like
an error message because no messages are found.  In reality, all the pending
messages have been handled and nothing has come in within the 500ms window.

Fixes: GRIF-67